+++ demo/src/demo/page/main/ws.cljs
@@ -3,18 +3,28 @@
    [taoensso.timbre :refer-macros [warn]]
    [reagent.core :as r]
    [re-frame.core :as rf]
+   [modular.ws.msg-handler :refer [-event-msg-handler]]
    [demo.helper.ui :refer [block2]]))
 
 ; WEBSOCKET
 (defn print-status [x]
   (warn "status: " x))
 
+(def time-a (r/atom "no time"))
+(def time2-a (r/atom "no time"))
+
+(defmethod -event-msg-handler :demo/time
+  [{:keys [?data]}]
+  (reset! time-a ?data))
+
+(defmethod -event-msg-handler :demo/time2
+  [{:keys [?data]}]
+  (reset! time2-a ?data))
+
 (defn demo-ws []
-  (let [t (rf/subscribe [:demo/time])
-        c (rf/subscribe [:ws/connected?])
-        t2  (rf/subscribe [:demo/time2])]
+  (let [c (rf/subscribe [:ws/connected?])]
     (fn []
       [block2 "websocket"
        [:p (str "connected:" (if @c @c "not connected"))]
-       [:p (str "time: " (if @t @t " no time received :-("))]
-       [:p "time as date: " (when @t2 (str @t2))]])))
+       [:p (str "time: "  @time-a)]
+       [:p (str "time as date: " @time2-a)]])))