+++ webly/src/deps.cljs
@@ -2,16 +2,12 @@
             "create-react-class" "^15.7.0"  ; is this needed?
             "react-dom-factories" "^1.0.2" ; is this needed?
             ; since reagent 1.1, reagent no longer brings react and react-dom
-            ;"react" "^17.0.1"
-            ;"react-dom" "^17.0.1"
             "react" "^18.0.0"
             "react-dom" "^18.0.0 "
             ; javascript unit tests 
             "karma" "^6.3.2"
             "karma-chrome-launcher" "^3.1.0"
             "karma-cljs-test" "^0.1.0"
-            ;"karma-junit-reporter"  "2.0.1"
-            ;"cross-env": "^6.0.3",
 
 ;
             }}